Rename table setup steps
